chore(cleanup): Phase 1b - Remove unused deps and apply Python idiom improvements
**Removed 6 unused dependencies:**
- pandas (not imported, ibis provides memtable functionality)
- repomix (not imported anywhere)
- mdformat (not imported anywhere)
- pydantic-settings (not imported anywhere)
- toml (not needed, Python 3.12+ has tomllib built-in)
- duckdb-engine (not imported anywhere)

**Applied 28 automated code improvements (ruff --fix --unsafe-fixes):**
- Error message formatting (EM101, EM102) - extract to variables
- Type parameter syntax updates (Python 3.12+ generic syntax)
- Whitespace cleanup and formatting fixes

**Fixed test API mismatches:**
- Updated test_privacy_pass.py to match current PrivacyConfig API
  - Old: anonymize_authors, pii_patterns, media_allowlist, enable_reidentification, reidentification_salt
  - New: tenant_id (required), detect_pii, allowed_media_domains, enable_reidentification_escrow, reidentification_retention_days
- Updated IR version from "v1" to "1.0.0" (14 occurrences)
- Fixed error message assertions to match actual error text
- Fixed import: PrivacyConfig now imported from egregora.privacy.config (not gate)

**Test Results:**
- Before: 24 failed, 451 passed
- After: 16 failed, 459 passed (fixed 8 privacy tests!)
- All 19 privacy tests now passing
- Remaining 16 failures are pre-existing issues in test_abstraction_layer.py and test_adapters.py (unrelated to Phase 1b changes)

**Files Modified:** 15 source files, 1 test file, pyproject.toml, uv.lock

Phase 1b complete - codebase cleanup quick wins (1 hour effort).

See CLEANUP_PHASE1_ENHANCED.md for full analysis.

+++ src/egregora/pipeline.py
@@ -204,12 +204,12 @@ def create_windows(  # noqa: PLR0913
                 max_hours = max_with_overlap.total_seconds() / 3600
                 if max_hours < 24:
                     # Use hours if < 1 day (floor to avoid exceeding)
-                    effective_step_size = int(math.floor(max_hours))
+                    effective_step_size = math.floor(max_hours)
                     effective_step_size = max(effective_step_size, 1)  # Minimum 1 hour
                     effective_step_unit = "hours"
                 else:
                     # Use days if >= 1 day (floor to avoid exceeding)
-                    effective_step_size = int(math.floor(max_with_overlap.days))
+                    effective_step_size = math.floor(max_with_overlap.days)
                     effective_step_size = max(effective_step_size, 1)  # Minimum 1 day
                     effective_step_unit = "days"
